BAEL-1756 Test typo fixed
